+++ packages/common/src/config/index.ts
@@ -1,2 +1,1 @@
 export { default as EnvironmentRegistry } from "./environmentRegistry";
-export { EnvironmentSchemaValidator } from "./environmentSchema";