+++ packages/common/src/constants.ts
@@ -1,0 +1,13 @@
+export const NodeEnvironment = {
+	DEV: "development",
+	PROD: "production",
+	TEST: "test"
+} as const;
+
+export const LogLevel = {
+	DEBUG: "debug",
+	ERROR: "error",
+	INFO: "info",
+	TRACE: "trace",
+	WARN: "warn"
+} as const;