+++ packages/common/src/index.ts
@@ -1,5 +1,7 @@
 export * from "./clients";
+export { EnvironmentRegistry } from "./config";
 export * from "./errors";
+export * from "./logger";
 export * from "./operations";
 export * from "./pipelines";
-export { ZodSchemaValidator } from "./validators";
+export { createEmailValidator, ZodSchemaValidator } from "./validators";