+++ packages/infrastructure-server/src/bootstrap.ts
@@ -1,6 +1,8 @@
+import { EnvironmentRegistry } from "@otuekong-portfolio/common";
+
 import { RateLimiterFactory, RedisClient, withRateLimitFactory } from "./redis";
 import { ResendClient } from "./resend";
-import { EnvironmentRegistry, SystemHealthClient, SystemHealthHandler } from "./server";
+import { SystemHealthClient, SystemHealthHandler } from "./server";
 import { ServerComponentMonitor } from "./types";
 
 /**