+++ packages/infrastructure-server/src/index.ts
@@ -6,5 +6,5 @@ export {
 	withRateLimit
 } from "./bootstrap";
 export { ResendClient } from "./resend";
-export { EnvironmentRegistry, type HttpResponse, type HttpRequest, type ResourceHandler } from "./server";
+export { type HttpResponse, type HttpRequest, type ResourceHandler } from "./server";
 export { type ServerComponentHealth } from "./types";