+++ packages/exhibit/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.1-rc-2] - 06/13/2026
+### Changed
+* Updating to pnpm v11.6.0
+
 ## [1.0.1-rc-1] - 06/07/2026
 ### Added
 * Adding and bootstrapping `HealthServiceHandler`